Tidied up code, rename index.js to component name.
